refactor(events): start re-organizing event definitions
- turn the event dependency ID type just a slice of IDs
- split code into multiple files (might still change) to organize
